requirements.txt
langchain
langchain-groq
langchain-community
tavily-python
pydantic
python-dotenv
